% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{video2pic_initialize}
\alias{video2pic_initialize}
\title{Initalize the python environment}
\usage{
video2pic_initialize(py_env = "video2pic")
}
\arguments{
\item{py_env}{the conda environment to use. "video2pic" by default}
}
\description{
Function to initalize the python environment
}
\examples{
video2pic(py_env = "video2pic")
}
