% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{video2pic_install}
\alias{video2pic_install}
\title{Install python dependencies for video2pic}
\usage{
video2pic_install(py_env = "video2pic")
}
\arguments{
\item{py_env}{the conda environment to use. "video2pic" by default}
}
\description{
Function to install all python dependencies for video2pic to work
}
\examples{
video2pic(py_env = "video2pic")
}
